refactor:user에서 auth로 변경
