+++ src/dtos/user.dto.js
@@ -42,18 +42,3 @@ export const responseFromPatchUserProfile = (data) =>{
         photo: data.photo
     }
 }
-//사용자 이메일 찾기 요청 DTP
-export const userForEmail = (body) => {
-    return{
-        name: body.name,
-        password: body.password
-    }
-}
-//사용자 이메일 찾기 응답 DTO
-export const responseFromUserEmail = (data) => {
-    return{
-        id: data.id,
-        name: data.name,
-        email: data.email
-    }
-}
