+++ src/errors/auth.error.js
@@ -130,3 +130,15 @@ export class SamePasswordError extends Error {
         this.statusCode = 400;
     }
 }
+
+//사용자가 입력한 정보와 일치한 이메일이 존재하지 않을 때,
+export class NoCorrectUserEmail extends Error {
+    errorCode = "A030";
+
+    constructor(reason,data) {
+        super(reason);
+        this.reason = reason;
+        this.statusCode = 400;
+        this.data = {userName: data};
+    }
+}
