+++ src/errors/user.error.js
@@ -22,15 +22,3 @@ export class NotFileAllowedError extends Error {
         this.allowed = allowed
     }
 }
-
-//사용자가 입력한 정보와 일치한 이메일이 존재하지 않을 때,
-export class NoCorrectUserEmail extends Error {
-    errorCode = "U030";
-
-    constructor(reason,data) {
-        super(reason);
-        this.reason = reason;
-        this.statusCode = 400;
-        this.data = {userName: data};
-    }
-}
